+++ src/components/user_card/user_card.vue
@@ -54,6 +54,18 @@
                 {{ user.name }}
               </div>
               <button
+                v-if="!isOtherUser && user.is_local"
+                class="button-unstyled edit-profile-button"
+                @click.stop="openProfileTab"
+              >
+                <FAIcon
+                  fixed-width
+                  class="icon"
+                  icon="edit"
+                  :title="$t('user_card.edit_profile')"
+                />
+              </button>
+              <button
                 v-if="isOtherUser && !user.is_local"
                 :href="user.statusnet_profile_url"
                 target="_blank"
@@ -426,7 +438,7 @@
     }
   }
 
-  .external-link-button {
+  .external-link-button, .edit-profile-button {
     cursor: pointer;
     width: 2.5em;
     text-align: center;
@@ -576,6 +588,10 @@
       margin: 0;
     }
   }
+}
+
+.sidebar .edit-profile-button {
+  display: none;
 }
 
 .user-counts {