+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added quick filters for notifications
 - Implemented user option to change sidebar position to the right side
 - Implemented user option to hide floating shout panel
+- Implemented "edit profile" button if viewing own profile which opens profile settings
 - Implemented user option to always show floating New Post button (normally mobile-only)
 
 ### Fixed